backend/pyproject.toml
[project]
name = "comfyui-task-manager-backend"
version = "0.1.0"
description = "FastAPI backend for ComfyUI task management"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "sqlalchemy==2.0.38",
  "greenlet==3.2.4",
  "asyncpg==0.30.0",
  "psycopg2-binary==2.9.10",
  "alembic==1.14.1",
  "pydantic-settings==2.7.1",
  "python-multipart==0.0.20",
  "httpx==0.28.1",
  "python-dateutil==2.9.0.post0",
  "websockets>=12.0,<14.0",
  "yt-dlp>=2025.10.14",
  "bcrypt>=4.0.0",
  "google-cloud-storage>=3.9.0",
]
